다시 시작 feat : 로또구매금액입력 함수( ui 함수 초기 구현)
