+++ src/App.js
@@ -1,23 +1,14 @@
 const MissionUtils = require("@woowacourse/mission-utils");
 const { Console } = require("@woowacourse/mission-utils");
 
-const { getLottoNumber, makeLottoArray } = require("./LottoGenerator");
-// const {
-//   makeWinLottoNumber,
-//   changePriceToCount,
-//   test,
-// } = require("./UiInputLogic");
-
-const UiInputLogic = require("./UiInputLogic");
-const Lotto = require("./Lotto");
+const LottoGameStart = require("./LottoGameStart");
 
 class App {
-  constructor() {
-    // this.lottoGame = new Lotto(this.lottoArray);
-    this.uiInputLogic = new UiInputLogic();
-  }
+  constructor() {}
 
-  play() {}
+  play() {
+    this.lottoGameStart = new LottoGameStart();
+  }
 }
 
 const app = new App();
