--- src/Message.js
@@ -1,18 +1,0 @@
-module.exports = Object.freeze({
-  INFORMATION: {
-    moneyInput: "구입금액을 입력해 주세요.",
-    lottoCount: "개를 구매했습니다.",
-    winningNumberInput: "당첨 번호를 입력해 주세요.",
-    bonusNumberInput: "보너스 번호를 입력해 주세요.",
-    winningStatistics: "당첨 통계\n---",
-    threeMatches: "3개 일치 (5,000원) -",
-    fourMatches: "4개 일치 (50,000원) -",
-    fiveMatches: "5개 일치 (1,500,000원) -",
-    fiveAndBonusMatches: "5개 일치, 보너스 볼 일치 (30,000,000원) -",
-    sixMatches: "6개 일치 (2,000,000,000원) -",
-  },
-  ERROR: {
-    lottoNumberValidWarning: "[ERROR] 로또 번호는 6개여야 합니다.",
-    moneyValidWarning: "[ERROR] 입력하신 금액이 1000원 단위가 아닙니다.",
-  },
-});
